Add Devops Release Item scripts (#1278)
- Add set of helpers to work with devops work items
- Add script to create devops release package items
- Update SemVer to support version type

@chidozieononiwu here is the helper scripts we need to use in the Prepare-Release scripts to create the release workitems.

fyi @czubair this is this script that we can start using to create the devops work items

Usage:

For existing package
```
.\eng\common\scripts\Update-DevOps-Release-WorkItem.ps1 "<language>" "<package_name>" "<package_version>" [-plannedDate 5/01/2021]
```
Language, package name, and version are all required. If the plannedDate is not set it will use today's date.

For new packages
```
.\eng\common\scripts\Update-DevOps-Release-WorkItem.ps1 "<language>" "<package_name>" "<package_version" -packageDisplayName "<package_display_name>" -serviceName "<service_name>" [-plannedDate 5/01/2021]
```
If the display name or service name is not passed for new packages the script will prompt the user to enter a value for them.

This script will create the package work item and link and/or create any parent items as necessary. It will also mark the package item as "In Release" and set the release type based on the version passed. It will also add the version to the list of planned releases.