Add lease to runs. (#1515)
This PR adds a PowerShell script that adds a retention lease to a pipeline run. Because you can only have one lease it first checks to see if a lease exists already, and if it does it removes it and adds it again for the specified validity period (there doesn't appear to be a patch operation according to the REST API docs).