Enable CI triggers on internal pipelines. (#1504)
This PR adds the CI trigger creation logic to the internal pipelines. It also fixes an error with the live test pipeline creation where it was hardcoded to use ```master``` instead of the context that was passed in on the command-line. Adding @danieljurek and @benbp for review on this one since it impacts live test pipelines.