[Test-Proxy] Fix default startup location in `docker-start-proxy.ps1`, Update `windows` tool initialize (#2290)
One can use script located at `eng/common/testproxy/docker-start-proxy.ps1` to start and stop a default docker configuration.

There is a second argument, `TargetFolder`, that is used to change the folder  will be bound to the volume of the new docker container.

The default used to be `"."`, so that the current folder would bind to the volume.

This PR changes the behavior so that if no target folder is provided, it will instead default to root of the containing repository. The ability to provide this context is unchanged.