Fail if the emulator doesn't start (#3509)
We had continueOnError set to true so if the emulator failed to start we would still try to run the tests which all fail but take a long time to run.  The timeout for this step is about 10mins of trying so that should be enough time and so we also don't really need another retry step.