Add tool link checks in nightly link checks (#6193)
We need to add a place to help catch broken links in things like eng/common so that we can fix them before they break more pipelines in the other languages. This will enable us to find them from the nightly run. Currently scoping to eng/common only because there are a number of broken links in the other part of the repo right now that I didn't have context to fix.