Add branch merging support to our mirror pipeline (#6392)
This adds the ability to do a branch merge as part of our mirroring
pipeline. The merge will allow configuration of which files to
move from source to target so it can be scoped to part of the branch.
Instead of a standard push it will create a PR with the merge so it
can be manually reviewed.