[FabricBot] Remove Rule Generator (#6512)
The focus of these changes is to remove the legacy tool that was once used
to generate Fabric Bot rules from the code owners files for a repository.  This
tool has been broken for a couple of years and worked only with the retired
FabricBot service API.

Now that FabricBot has been retired, there is no need for retaining this tool.